+++ misc/tests/algorithm_test.c
@@ -72,12 +72,12 @@ TEST(algorithm, c_filter)
     EXPECT_EQ(42, sum);
 
     uint64_t hash = 0;
-    c_filter(IVec, vec, hash = chash_mix(hash, *value));
-    EXPECT_EQ(658, hash);
+    c_filter(IVec, vec, hash = chash_mix(hash, (uint64_t)*value));
+    EXPECT_EQ(658, (isize)hash);
 
     hash = 0;
-    c_filter_reverse(IVec, vec, hash = chash_mix(hash, *value));
-    EXPECT_EQ(10897, hash);
+    c_filter_reverse(IVec, vec, hash = chash_mix(hash, (uint64_t)*value));
+    EXPECT_EQ(10897, (isize)hash);
 
     sum = 0;
     c_filter(IVec, vec, sum += *value);
